refactored to remove exported gnatsd dependency.
