+++ test/enc_test.go
@@ -1,6 +1,6 @@
-// Copyright 2012-2015 Apcera Inc. All rights reserved.
+// Copyright 2012-2017 Apcera Inc. All rights reserved.
 
-package builtin_test
+package test
 
 import (
 	"bytes"
@@ -9,13 +9,12 @@ import (
 
 	"github.com/nats-io/go-nats"
 	"github.com/nats-io/go-nats/encoders/builtin"
-	"github.com/nats-io/go-nats/test"
 )
 
 const TEST_PORT = 8168
 
-func NewEConn(t *testing.T) *nats.EncodedConn {
-	ec, err := nats.NewEncodedConn(test.NewConnection(t, TEST_PORT), nats.DEFAULT_ENCODER)
+func NewDefaultEConn(t *testing.T) *nats.EncodedConn {
+	ec, err := nats.NewEncodedConn(NewConnection(t, TEST_PORT), nats.DEFAULT_ENCODER)
 	if err != nil {
 		t.Fatalf("Failed to create an encoded connection: %v\n", err)
 	}
@@ -23,10 +22,10 @@ func NewEConn(t *testing.T) *nats.EncodedConn {
 }
 
 func TestConstructorErrs(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
-	c := test.NewConnection(t, TEST_PORT)
+	c := NewConnection(t, TEST_PORT)
 	_, err := nats.NewEncodedConn(nil, "default")
 	if err == nil {
 		t.Fatal("Expected err for nil connection")
@@ -44,10 +43,10 @@ func TestConstructorErrs(t *testing.T) {
 }
 
 func TestMarshalString(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
-	ec := NewEConn(t)
+	ec := NewDefaultEConn(t)
 	defer ec.Close()
 	ch := make(chan bool)
 
@@ -60,7 +59,7 @@ func TestMarshalString(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish("enc_string", testString)
-	if e := test.Wait(ch); e != nil {
+	if e := Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -69,10 +68,10 @@ func TestMarshalString(t *testing.T) {
 }
 
 func TestMarshalBytes(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
-	ec := NewEConn(t)
+	ec := NewDefaultEConn(t)
 	defer ec.Close()
 	ch := make(chan bool)
 
@@ -85,7 +84,7 @@ func TestMarshalBytes(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish("enc_bytes", testBytes)
-	if e := test.Wait(ch); e != nil {
+	if e := Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -94,10 +93,10 @@ func TestMarshalBytes(t *testing.T) {
 }
 
 func TestMarshalInt(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
-	ec := NewEConn(t)
+	ec := NewDefaultEConn(t)
 	defer ec.Close()
 	ch := make(chan bool)
 
@@ -110,7 +109,7 @@ func TestMarshalInt(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish("enc_int", testN)
-	if e := test.Wait(ch); e != nil {
+	if e := Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -119,10 +118,10 @@ func TestMarshalInt(t *testing.T) {
 }
 
 func TestMarshalInt32(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
-	ec := NewEConn(t)
+	ec := NewDefaultEConn(t)
 	defer ec.Close()
 	ch := make(chan bool)
 
@@ -135,7 +134,7 @@ func TestMarshalInt32(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish("enc_int", testN)
-	if e := test.Wait(ch); e != nil {
+	if e := Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -144,10 +143,10 @@ func TestMarshalInt32(t *testing.T) {
 }
 
 func TestMarshalInt64(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
-	ec := NewEConn(t)
+	ec := NewDefaultEConn(t)
 	defer ec.Close()
 	ch := make(chan bool)
 
@@ -160,7 +159,7 @@ func TestMarshalInt64(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish("enc_int", testN)
-	if e := test.Wait(ch); e != nil {
+	if e := Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -169,10 +168,10 @@ func TestMarshalInt64(t *testing.T) {
 }
 
 func TestMarshalFloat32(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
-	ec := NewEConn(t)
+	ec := NewDefaultEConn(t)
 	defer ec.Close()
 	ch := make(chan bool)
 
@@ -185,7 +184,7 @@ func TestMarshalFloat32(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish("enc_float", testN)
-	if e := test.Wait(ch); e != nil {
+	if e := Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -194,10 +193,10 @@ func TestMarshalFloat32(t *testing.T) {
 }
 
 func TestMarshalFloat64(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
-	ec := NewEConn(t)
+	ec := NewDefaultEConn(t)
 	defer ec.Close()
 	ch := make(chan bool)
 
@@ -210,7 +209,7 @@ func TestMarshalFloat64(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish("enc_float", testN)
-	if e := test.Wait(ch); e != nil {
+	if e := Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -219,10 +218,10 @@ func TestMarshalFloat64(t *testing.T) {
 }
 
 func TestMarshalBool(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
-	ec := NewEConn(t)
+	ec := NewDefaultEConn(t)
 	defer ec.Close()
 	ch := make(chan bool)
 	expected := make(chan bool, 1)
@@ -237,7 +236,7 @@ func TestMarshalBool(t *testing.T) {
 
 	expected <- false
 	ec.Publish("enc_bool", false)
-	if e := test.Wait(ch); e != nil {
+	if e := Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -246,7 +245,7 @@ func TestMarshalBool(t *testing.T) {
 
 	expected <- true
 	ec.Publish("enc_bool", true)
-	if e := test.Wait(ch); e != nil {
+	if e := Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -255,10 +254,10 @@ func TestMarshalBool(t *testing.T) {
 }
 
 func TestExtendedSubscribeCB(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
-	ec := NewEConn(t)
+	ec := NewDefaultEConn(t)
 	defer ec.Close()
 
 	ch := make(chan bool)
@@ -276,7 +275,7 @@ func TestExtendedSubscribeCB(t *testing.T) {
 		ch <- true
 	})
 	ec.Publish(subject, testString)
-	if e := test.Wait(ch); e != nil {
+	if e := Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -285,10 +284,10 @@ func TestExtendedSubscribeCB(t *testing.T) {
 }
 
 func TestExtendedSubscribeCB2(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
-	ec := NewEConn(t)
+	ec := NewDefaultEConn(t)
 	defer ec.Close()
 
 	ch := make(chan bool)
@@ -310,7 +309,7 @@ func TestExtendedSubscribeCB2(t *testing.T) {
 		ch <- true
 	})
 	ec.PublishRequest(oSubj, oReply, testString)
-	if e := test.Wait(ch); e != nil {
+	if e := Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -319,10 +318,10 @@ func TestExtendedSubscribeCB2(t *testing.T) {
 }
 
 func TestRawMsgSubscribeCB(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
-	ec := NewEConn(t)
+	ec := NewDefaultEConn(t)
 	defer ec.Close()
 
 	ch := make(chan bool)
@@ -345,7 +344,7 @@ func TestRawMsgSubscribeCB(t *testing.T) {
 		ch <- true
 	})
 	ec.PublishRequest(oSubj, oReply, testString)
-	if e := test.Wait(ch); e != nil {
+	if e := Wait(ch); e != nil {
 		if ec.LastError() != nil {
 			e = ec.LastError()
 		}
@@ -354,10 +353,10 @@ func TestRawMsgSubscribeCB(t *testing.T) {
 }
 
 func TestEncRequest(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
-	ec := NewEConn(t)
+	ec := NewDefaultEConn(t)
 	defer ec.Close()
 
 	expectedResp := "I can help!"
@@ -378,10 +377,10 @@ func TestEncRequest(t *testing.T) {
 }
 
 func TestEncRequestReceivesMsg(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
-	ec := NewEConn(t)
+	ec := NewDefaultEConn(t)
 	defer ec.Close()
 
 	expectedResp := "I can help!"
@@ -402,10 +401,10 @@ func TestEncRequestReceivesMsg(t *testing.T) {
 }
 
 func TestAsyncMarshalErr(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
-	ec := NewEConn(t)
+	ec := NewDefaultEConn(t)
 	defer ec.Close()
 
 	ch := make(chan bool)
@@ -422,7 +421,7 @@ func TestAsyncMarshalErr(t *testing.T) {
 	}
 
 	ec.Publish(subject, testString)
-	if e := test.Wait(ch); e != nil {
+	if e := Wait(ch); e != nil {
 		t.Fatalf("Did not receive the message: %s", e)
 	}
 }
