+++ test/protobuf_test.go
@@ -1,4 +1,4 @@
-package protobuf_test
+package test
 
 import (
 	"reflect"
@@ -6,16 +6,15 @@ import (
 	"time"
 
 	"github.com/nats-io/go-nats"
-	"github.com/nats-io/go-nats/test"
 
 	"github.com/nats-io/go-nats/encoders/protobuf"
 	pb "github.com/nats-io/go-nats/encoders/protobuf/testdata"
 )
 
-const TEST_PORT = 8068
+// const TEST_PORT = 8068
 
-func NewProtoEncodedConn(tl test.TestLogger) *nats.EncodedConn {
-	ec, err := nats.NewEncodedConn(test.NewConnection(tl, TEST_PORT), protobuf.PROTOBUF_ENCODER)
+func NewProtoEncodedConn(tl TestLogger) *nats.EncodedConn {
+	ec, err := nats.NewEncodedConn(NewConnection(tl, TEST_PORT), protobuf.PROTOBUF_ENCODER)
 	if err != nil {
 		tl.Fatalf("Failed to create an encoded connection: %v\n", err)
 	}
@@ -23,7 +22,7 @@ func NewProtoEncodedConn(tl test.TestLogger) *nats.EncodedConn {
 }
 
 func TestProtoMarshalStruct(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
 	ec := NewProtoEncodedConn(t)
@@ -44,13 +43,13 @@ func TestProtoMarshalStruct(t *testing.T) {
 	})
 
 	ec.Publish("protobuf_test", me)
-	if e := test.Wait(ch); e != nil {
+	if e := Wait(ch); e != nil {
 		t.Fatal("Did not receive the message")
 	}
 }
 
 func TestProtoNilRequest(t *testing.T) {
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
 	ec := NewProtoEncodedConn(t)
@@ -97,7 +96,7 @@ func BenchmarkPublishProtobufStruct(b *testing.B) {
 	// stop benchmark for set-up
 	b.StopTimer()
 
-	s := test.RunServerOnPort(TEST_PORT)
+	s := RunServerOnPort(TEST_PORT)
 	defer s.Shutdown()
 
 	ec := NewProtoEncodedConn(b)
@@ -122,7 +121,7 @@ func BenchmarkPublishProtobufStruct(b *testing.B) {
 
 	for n := 0; n < b.N; n++ {
 		ec.Publish("protobuf_test", me)
-		if e := test.Wait(ch); e != nil {
+		if e := Wait(ch); e != nil {
 			b.Fatal("Did not receive the message")
 		}
 	}
